Rename mod to native_staking_callback proper
