+++ contracts/native-staking/src/native_staking_callback.rs
